Atualização Módulo01

+++ Modulo01/Pontos2D-C/Ponto2D.c
@@ -1,9 +1,9 @@
 #include <math.h>
 #include "Ponto2D.h"
 
-float distancia(Ponto2D p1, Ponto2D p2) {
-    float dx = p1.x - p2.x;
-    float dy = p1.y - p2.y;
-    return ((float)sqrt (dx * dx + dy * dy));
+double distancia(Ponto2D p1, Ponto2D p2) {
+    double dx = p1.x - p2.x;
+    double dy = p1.y - p2.y;
+    return sqrt (dx * dx + dy * dy);
 }
