+++ Modulo01/Pontos2D-C/Ponto2D.h
@@ -2,10 +2,10 @@
 #define PONTO2D_H
 
 typedef struct Ponto2D {
-    float x, y;
+    double x, y;
 } Ponto2D;
 
-float distancia(Ponto2D p1, Ponto2D p2);
+double distancia(Ponto2D p1, Ponto2D p2);
 
 #endif /* PONTO2D_H */
